#53: refactor and implemented PipeSet random reposition
